[#5070] [#5201] Ensure WebSocket*Handshaker removes correct handler during handshake
Motivation:

When a user has multiple instances of Http codecs in the pipline the WebSocket*Handshaker may remove the wrong ones as it will just remove the first that are found in the pipeline. We should better provide a way to have it search the "nearest" codec and remove this one.

Modifications:

Add a new paramater which is a Set of ChannelHandlers to remove during / after handshake.

Result:

Safe removal of handlers.